full laag feed refactor for groups
