+++ components/laags/laag-feed/comment-card.tsx
@@ -1,0 +1,152 @@
+"use client"
+
+import { useState } from "react"
+import { createClient } from "@/utils/supabase/client"
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
+import { useAvatar } from "@/hooks/useAvatar"
+import { format } from "date-fns"
+import { Button } from "@/components/ui/button"
+import { Clock, MoreHorizontal, Trash2, Pencil } from "lucide-react"
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
+import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger } from "@/components/ui/alert-dialog"
+import { toast } from "sonner"
+import { CommentCardProps } from "@/types"
+
+export function CommentCard({ comment, onDelete }: CommentCardProps) {
+  const [isEditing, setIsEditing] = useState(false)
+  const [editedComment, setEditedComment] = useState(comment.comment)
+  const [isSubmitting, setIsSubmitting] = useState(false)
+  const supabase = createClient()
+  const userAvatarUrl = useAvatar(comment.user?.avatar_url || null)
+
+  const handleUpdate = async () => {
+    if (!editedComment.trim() || editedComment === comment.comment) {
+      setIsEditing(false)
+      return
+    }
+
+    setIsSubmitting(true)
+    try {
+      const { error } = await supabase
+        .from("comments")
+        .update({ comment: editedComment.trim() })
+        .eq("id", comment.id)
+
+      if (error) throw error
+
+      toast.success("Comment updated successfully")
+      setIsEditing(false)
+      onDelete()
+    } catch (error) {
+      console.error("Error updating comment:", error)
+      toast.error("Failed to update comment")
+    } finally {
+      setIsSubmitting(false)
+    }
+  }
+
+  const handleDelete = async () => {
+    try {
+      const { error } = await supabase
+        .from("comments")
+        .update({ is_deleted: true })
+        .eq("id", comment.id)
+
+      if (error) throw error
+
+      toast.success("Comment deleted successfully")
+      onDelete()
+    } catch (error) {
+      console.error("Error deleting comment:", error)
+      toast.error("Failed to delete comment")
+    }
+  }
+
+  if (!comment.user) return null
+
+  return (
+    <div className="flex gap-3">
+      <Avatar className="h-8 w-8 border">
+        <AvatarImage src={userAvatarUrl || undefined} />
+        <AvatarFallback>{comment.user.full_name.charAt(0)}</AvatarFallback>
+      </Avatar>
+
+      <div className="flex-1 space-y-2">
+        <div className="flex items-start justify-between">
+          <div>
+            <p className="text-sm font-medium">{comment.user.full_name}</p>
+            <div className="flex items-center gap-1 text-xs text-muted-foreground">
+              <Clock className="h-3 w-3" />
+              <span>{format(new Date(comment.created_at), "MMM d, yyyy 'at' h:mm a")}</span>
+            </div>
+          </div>
+
+          <DropdownMenu>
+            <DropdownMenuTrigger asChild>
+              <Button variant="ghost" size="icon" className="h-8 w-8">
+                <MoreHorizontal className="h-4 w-4" />
+                <span className="sr-only">More options</span>
+              </Button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent align="end">
+              <DropdownMenuItem onClick={() => setIsEditing(true)}>
+                <Pencil className="mr-2 h-4 w-4" />
+                Edit
+              </DropdownMenuItem>
+              <AlertDialog>
+                <AlertDialogTrigger asChild>
+                  <DropdownMenuItem onSelect={(e) => e.preventDefault()}>
+                    <Trash2 className="mr-2 h-4 w-4 text-destructive" />
+                    Delete
+                  </DropdownMenuItem>
+                </AlertDialogTrigger>
+                <AlertDialogContent>
+                  <AlertDialogHeader>
+                    <AlertDialogTitle>Delete Comment</AlertDialogTitle>
+                    <AlertDialogDescription>
+                      Are you sure you want to delete this comment? This action cannot be undone.
+                    </AlertDialogDescription>
+                  </AlertDialogHeader>
+                  <AlertDialogFooter>
+                    <AlertDialogCancel>Cancel</AlertDialogCancel>
+                    <AlertDialogAction onClick={handleDelete} className="bg-destructive text-destructive-foreground hover:bg-destructive/90">
+                      Delete
+                    </AlertDialogAction>
+                  </AlertDialogFooter>
+                </AlertDialogContent>
+              </AlertDialog>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        </div>
+
+        {isEditing ? (
+          <div className="space-y-2">
+            <textarea
+              value={editedComment}
+              onChange={(e) => setEditedComment(e.target.value)}
+              className="w-full p-2 text-sm border rounded-md"
+              rows={2}
+            />
+            <div className="flex justify-end gap-2">
+              <Button
+                variant="outline"
+                size="sm"
+                onClick={() => {
+                  setIsEditing(false)
+                  setEditedComment(comment.comment)
+                }}
+              >
+                Cancel
+              </Button>
+              <Button size="sm" onClick={handleUpdate} disabled={isSubmitting}>
+                {isSubmitting ? "Updating..." : "Update"}
+              </Button>
+            </div>
+          </div>
+        ) : (
+          <p className="text-sm whitespace-pre-wrap">{comment.comment}</p>
+        )}
+      </div>
+    </div>
+  )
+}
